feat: 주석 정리 & Tree shaking 활성화
